Move balance.ts to api/utils directory
#175
